Costumizando a interface, adicionando locais para o usuário definir as propriedades do produto.

+++ Compras/Models/Produto.cs
@@ -8,6 +8,6 @@ namespace Compras.Models
         public int Id { get; set; }
         public string Descricao { get; set; }
         public double Quantidade { get; set; }
-        public decimal Preco { get; set; }
+        public double Preco { get; set; }
     }
 }
